@@ -3,15 +3,18 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
+		"@emotion/react": "^11.7.1",
+		"@emotion/styled": "^11.6.0",
+		"@mui/icons-material": "^5.2.5",
 		"@mui/material": "^5.2.5",
-		"@testing-library/jest-dom": "^4.2.4",
+		"@testing-library/jest-dom": "^5.16.1",
 		"@testing-library/react": "^9.3.2",
 		"@testing-library/user-event": "^7.1.2",
 		"axios": "^0.24.0",
-		"react": "^16.13.1",
+		"react": "^16.14.0",
 		"react-dom": "^16.13.1",
 		"react-google-login": "^5.1.10",
-		"react-scripts": "3.4.1"
+		"react-scripts": "^5.0.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
